updated styling for legend and load balance button
