+++ client/scss/Selector.scss
@@ -41,21 +41,6 @@
   font-family: $buttonTitle;
   align-content: center;
 }
-.loadBalanceBtn {
-  margin-top: 10px;
-  border-radius: 4px;
-  background-color: $saamsaYellow;
-  border: 1px solid black;
-  height: 30px;
-  font-size: 15px;
-  font-family: $buttonTitle;
-  color: $saamsaBlack;
-  width: 200px;
-  position: absolute;
-  top: 557px;
-  right: 70px;
-  z-index: 1;
-}
 
 // #heading, #loggedIn{
 //         display:inline;
